requirements.txt
amqp==5.1
boto3==1.24
botocore==1.27
celery==5.2
charset-normalizer==2.0.0
Django==3.2.14
django-bootstrap-form==3.4
django-crispy-forms==1.14.0
django-filter==22.1
django-leaflet==0.28.3
django-storages==1.12.3
django-user-visit==0.5.1
kombu>=5.2.1
requests==2.28
s3transfer==0.6
six==1.16
sorl-thumbnail==12.8
sqlparse==0.4
redis
pillow
gunicorn==20.1.0
django-heroku==0.3.1
django-heroku-memcacheify
django-cors-headers
django-import-export==2.9.0
gevent
gevent-socketio
gevent-websocket
django-db-geventpool